+++ DellOpenManage/Classes/ConfigurationCompliance.psm1
@@ -1,0 +1,8 @@
+Class ConfigurationCompliance {
+    [Int]$Id
+    [String]$ServiceTag
+    [String]$DeviceModel
+    [String]$DeviceName
+    [String]$ComplianceStatus
+    [String]$InventoryTime
+}