+++ DellOpenManage/Public/OME/New-OMETemplateFromDevice.ps1
@@ -30,9 +30,9 @@ limitations under the License.
 .PARAMETER Component
     Components to include in the template (Default="All", "iDRAC", "BIOS", "System", "NIC", "LifecycleController", "RAID", "EventFilters")
 .PARAMETER Type
-    Type of template to create (Default="Deployment", "Compliance")
+    Type of template to create (Default="Deployment", "Configuration")
     Deployment: Only 1 template assigned to a device, used with Virtual Identities
-    Compliance: Many templates can be assigned to a device, used with Configuration Compliance
+    Configuration: Many templates can be assigned to a device, used with Configuration Compliance
 .PARAMETER Wait
     Wait for job to complete
 .PARAMETER WaitTime
@@ -43,8 +43,8 @@ limitations under the License.
     New-OMETemplateFromDevice -Component "iDRAC", "BIOS" -Device $("37KP0ZZ" | Get-OMEDevice -FilterBy "ServiceTag") -Wait
     Create new deployment template from device
 .EXAMPLE
-    New-OMETemplateFromDevice -TemplateType "Compliance" -Component "iDRAC", "BIOS" -Device $("37KP0ZZ" | Get-OMEDevice -FilterBy "ServiceTag") -Wait
-    Create new compliance template from device
+    New-OMETemplateFromDevice -TemplateType "Configuration" -Component "iDRAC", "BIOS" -Device $("37KP0ZZ" | Get-OMEDevice -FilterBy "ServiceTag") -Wait
+    Create new configuration template from device
 #>
 
 [CmdletBinding()]
@@ -63,7 +63,7 @@ param(
     [String[]]$Component = @("All"),
 
     [Parameter(Mandatory=$false)]
-    [ValidateSet("Deployment", "Compliance")]
+    [ValidateSet("Deployment", "Configuration")]
     [String]$TemplateType = "Deployment",
 
     [Parameter(Mandatory=$false)]
@@ -88,7 +88,7 @@ Process {
         $Headers     = @{}
         $Headers."X-Auth-Token" = $SessionAuth.Token
         $TEMPLATE_TYPE_MAP = @{
-            "Compliance" = 1;
+            "Configuration" = 1;
             "Deployment" = 2
         }
         $TemplateUrl = $BaseUri + "/api/TemplateService/Templates"
@@ -120,9 +120,7 @@ Process {
         return $TemplateId
     }
     Catch {
-        Write-Error ($_.ErrorDetails)
-        Write-Error ($_.Exception | Format-List -Force | Out-String)
-        Write-Error ($_.InvocationInfo | Format-List -Force | Out-String)
+        Resolve-Error $_
     }
 }