+++ Tests/Device.Tests.ps1
@@ -2,7 +2,7 @@ $credentials = New-Object -TypeName System.Management.Automation.PSCredential -A
 Connect-OMEServer -Name $Global:OMEServer -Credentials $credentials -IgnoreCertificateWarning
 Describe "Device Tests" {
     BeforeEach {
-        $TestDeviceHosts = @("100.79.6.22", "100.79.6.63")
+        $TestDeviceHosts = @("100.79.6.22", "100.79.6.27", "100.79.6.63")
         $TestDeviceHostsNew = @("100.79.6.64")
         $TestDeviceServiceTags = @("37KP0Q2")
         $TestDiscoveryJobName = "TestDiscovery01"
@@ -20,19 +20,19 @@ Describe "Device Tests" {
             $TestDiscoveryJobName | Get-OMEDiscovery | Measure-Object | Select-Object -ExpandProperty Count | Should -Be 1
         }
 
-        It "EditDiscoveryJobAppend > Should return 2 targets hosts" {
+        It "EditDiscoveryJobAppend > Should return 4 targets hosts" {
             $TestDiscoveryJobName | Get-OMEDiscovery | Edit-OMEDiscovery -Hosts $TestDeviceHostsNew -Mode "Append" -DiscoveryUserName $TestiDRACUsername -DiscoveryPassword $(ConvertTo-SecureString $TestiDRACPassword -AsPlainText -Force)
+            $TestDiscoveryJobName | Get-OMEDiscovery | Select-Object -ExpandProperty Hosts | Measure-Object | Select-Object -ExpandProperty Count | Should -Be 4
+        }
+
+        It "EditDiscoveryJobRemove > Should return 3 target hosts" {
+            $TestDiscoveryJobName | Get-OMEDiscovery | Edit-OMEDiscovery -Hosts $TestDeviceHostsNew -Mode "Remove" -DiscoveryUserName $TestiDRACUsername -DiscoveryPassword $(ConvertTo-SecureString $TestiDRACPassword -AsPlainText -Force)
             $TestDiscoveryJobName | Get-OMEDiscovery | Select-Object -ExpandProperty Hosts | Measure-Object | Select-Object -ExpandProperty Count | Should -Be 3
         }
 
-        It "EditDiscoveryJobRemove > Should return 1 target hosts" {
-            $TestDiscoveryJobName | Get-OMEDiscovery | Edit-OMEDiscovery -Hosts $TestDeviceHostsNew -Mode "Remove" -DiscoveryUserName $TestiDRACUsername -DiscoveryPassword $(ConvertTo-SecureString $TestiDRACPassword -AsPlainText -Force)
-            $TestDiscoveryJobName | Get-OMEDiscovery | Select-Object -ExpandProperty Hosts | Measure-Object | Select-Object -ExpandProperty Count | Should -Be 2
-        }
-
-        It "EditDiscoveryJobKeepHosts > Should return 1 target hosts" {
+        It "EditDiscoveryJobKeepHosts > Should return 3 target hosts" {
             $TestDiscoveryJobName | Get-OMEDiscovery | Edit-OMEDiscovery -DiscoveryUserName $TestiDRACUsername -DiscoveryPassword $(ConvertTo-SecureString $TestiDRACPassword -AsPlainText -Force)
-            $TestDiscoveryJobName | Get-OMEDiscovery | Select-Object -ExpandProperty Hosts | Measure-Object | Select-Object -ExpandProperty Count | Should -Be 2
+            $TestDiscoveryJobName | Get-OMEDiscovery | Select-Object -ExpandProperty Hosts | Measure-Object | Select-Object -ExpandProperty Count | Should -Be 3
         }
 
         It "EditDiscoveryJobRunLater > Should return matching cron string" {